// src/app/interfaces/common/conference.interface.ts
export interface Conference {
  _id?: string;
  name?: string;
  nameEn?: string;
  image?: string;
  title?: string;
  seoImage?: string;
  titleEn?: string;
  description?: string;
  isHtml?: boolean;
  descriptionEn?: string;
  isHtmlEn?: boolean;
  url?: string;
  createdAt?: Date;
  updatedAt?: Date;
}